#refactor

+++ WebApiCarProject/Infrastructure/Entities/BaseEntity.cs
@@ -5,7 +5,7 @@ namespace WebApiCarProject.Infrastructure.Entities;
 public class BaseEntity
 {
     [Key] 
-    public long Id { get; init; }
+    public int Id { get; init; }
 
     public DateTime CreatedAtUtc { get; init; } = DateTime.UtcNow;
 }
